Move image load func to imgutil package
